* Adding KBDiagonalWrapStrategy (bounce will follow), which controls   the kbdiagonalshotstrategy
btw. don't think the AI is stupid, if you always see that only
ONE strategy is used in ONE game, the point is that those Strategies
only change if you start a new game! Not in the game, _yet_
Kevin has promised to changed that in a day or two

svn path=/trunk/kdegames/kbattleship/; revision=95144