Avoid using too much memory when reading the receiving data
When the server receives something from the internet, it must be careful
to not read everything they will throw to it.
for example, an xml with valid tags but with gigantic data.

Before, a modified xml with X*1,000,000 'a' in the name field, will make
the server allocate XGb or more.
After, at most 4Kb will be used.
The largest package will contain all the ships positions, and it uses
less than 1Kb.