implement the possibility of non adjacent ships
Before, the only game option was to play with adjacent ships.
Now it is possible to choose between the two options, adjacent or not,
including the network games.
The game mode can not be changed in the middle of a battle.
The local user settings are not affected by a network game. Both
players will play the same game type, but after that, the local options
will apply for new local games.

FEATURE: 168659
FIXED-IN: 11.0
REVIEW: 106772
GUI: There is a new settings menu entry (Adjacent ships)