utilities for the several ships patch
* Implement a bool canAddShipOfSize(size). When a ship of size size
can not be placed, then the AI or the player should start again. The
other option is to implement an Undo.
* Implement a clear() method to be able to start again
* In the battleField also add another hidden board with the cells having
only two status (free, busy) to be able to know the availability of the
cells.
It is easier for me in this way than the other way (sending a signal
to clear the ship borders when the shooting starts).

Also, done Roney suggestions.

Moved from reviewboard to branch

https://git.reviewboard.kde.org/r/114845/