Two phase place ships, enable restart placing ships
There are now two phases in the game, also used in the seaview status,
placing and playing. Placing can be restarted several times without
disturbing the other player.
Before:
  controller called entity start and waited for 2 ready() signals.
Now:
  controller calls entity startPlacing and waits for 2 ready() signals,
then it calls start and waits for 2 more answers.
If a player can not finish placing its ships, a restartPlacingShips()
signal is emited (and not the ready() signal). It reaches playfield
where a messagebox is displayed to restart or abort the game
(missing, I'm not sure about this option).
 This patch only affects the human player (the AI, so fast, will retry
automatically until a good placement is reached. It will be addressed
in the AI placing patch).
 For the network player, it assumes the other player has finished placing
its ships, but still needs another message (received when the other
player finishes placing ships) to start shooting.