remove my TODOs, FIXME and some warnings
There is still something to test: an old
client (pre-several ships) vs. new client
(as a server and as client).
Old as a server, the new client should
not be able to play with several ships.
Old as client, the server should not
be able to play with several ships.